feat(enforce-post): Add rule for enforcing the use of HTTP POST method when sending operations to the upstream API (#21)
feat(enforce-post): Add rule for enforcing the use of HTTP POST method
when sending operations to the upstream API

When enabled, enforces the use of HTTP POST method for sending
operations to the GraphQL API upstream.

Still allows other methods to reach the upstream API, as long as they do
not contain an operation. This is useful for i.e. GraphiQL to be
reached.

---------

Co-authored-by: ldebruijn <ldebruijn@bol.com>