Fix refresh ticker dropping operations (#74)
We found that the remote refreshing of persisted operations would
sometimes drop operations from specific files. There seems to be a
scenario where a file is correctly downloaded from the remote, and
stored locally, but when reading the file right after it would see an
empty file still.

This cleans up the code around that mechanism, provides additional
protections for concurrent processing and fixes the issue. The applied
fix may not be the prettiest, but it does the trick and can be supplied
faster than diving into why there is a discrepancy when writing to a
file and reading it straight after.

---------

Co-authored-by: ldebruijn <ldebruijn@bol.com>