Rewrite persisted ops loader (#114)
Our initial strategy for loading persisted operations was flawed, mainly
because of trying to use the local filesystem as a state store,
requiring it to be consistent immediately after loading from a remote,
and trying to read from local file storage right after.

This design was chosen in order to decouple local state from remote
loading strategies, but this can be achieved in other ways.

This PR rewrites the loading behavior to use just the internal state in
memory, and a loading strategy to append to it. This removes the
dependency on the local file system for persisting this state, and
simplifies the interactions for loading persisted operations.

Now, a single `loader` can be specified which currently are `local`,
meaning files are read from the local file system. `gcp` meaning files
are downloaded from a GCP bucket and loaded straight into the internal
state, or `noop` meaning no files are loaded.

It also updates a number of metrics as well as the accompanying
documentation

---------

Co-authored-by: ldebruijn <ldebruijn@bol.com>