Limit request bytes (#138)
Add an option to limit the request body size to prevent OOM attacks
through excessively large request bodies.
This option is enabled by default with a size limit of 100kb. While this
is small for a request, many GraphQL requests aren't that large. This
size was chosen to allow for sufficiently complex queries, combined with
some potential use of batching and other GraphQL features, while also
providing sufficient protection.