feat: add async option to access logging (#212)
This PR solves the following issue:

After adding more spans to the tracing output, we noticed that when we
noticed p99 latency issues on our GraphQL endpoint, the most time was
spent on the access-logging part of the request, the other validations
were pretty fast.

Reading up on how slog works with stdout, is that it contains a lock,
and also os.Stdout is having a fixed buffer size, and underlying systems
like the Linux kernel and perhaps even the container-runtime all also
have possibly limitations on how much you can throw at stdout.

To fix this buffering, and remove it from the hot path, in this PR a new
buffer (no pun intended) is created, so it can throw the request onto
that buffer, and move forward with processing the rest of the request. A
separate goroutine is then responsible for actually writing all the
buffered entries out to the logger again, which then is allowed to take
as long as it needs.

Some metrics are added to see how full the buffers are and whether we
still need to drop logs or not.